Update ESLint to newest version, add the missing plugins, and update the code to conform to new rules
