+++ common/package.json
@@ -12,13 +12,10 @@
   "devDependencies": {
     "@babel/preset-typescript": "^7.23.3",
     "@eslint/js": "^9.4.0",
-    "@types/eslint__js": "^8.42.3",
     "@types/node": "^20.11.28",
     "@vitest/coverage-v8": "^3.1.1",
-    "eslint": "^8.57.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.12.0",
     "vitest": "^3.1.1"
   }
 }
