+++ eslint.config.mjs
@@ -1,18 +1,16 @@
 import globals from 'globals';
 import pluginJs from '@eslint/js';
 import tseslint from 'typescript-eslint';
-import pluginReactConfig from 'eslint-plugin-react/configs/recommended.js';
-import { fixupConfigRules } from '@eslint/compat';
+import pluginReact from 'eslint-plugin-react';
+import reactHooks from 'eslint-plugin-react-hooks';
+import reactRefresh from 'eslint-plugin-react-refresh';
 
 export default [
   { languageOptions: { globals: { ...globals.browser, ...globals.node } } },
   pluginJs.configs.recommended,
   ...tseslint.configs.recommended,
-  {
-    files: ['**/*.jsx'],
-    languageOptions: { parserOptions: { ecmaFeatures: { jsx: true } } },
-  },
-  ...fixupConfigRules(pluginReactConfig),
+  pluginReact.configs.flat.recommended,
+  pluginReact.configs.flat['jsx-runtime'],
   {
     rules: {
       'react/react-in-jsx-scope': 'off',
@@ -26,4 +24,6 @@ export default [
   {
     ignores: ['**/dist/', '**/coverage/'],
   },
+  reactHooks.configs['recommended-latest'],
+  reactRefresh.configs.recommended
 ];
