+++ frontend/package.json
@@ -25,6 +25,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.51.1",
+    "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
@@ -32,14 +33,9 @@
     "@types/node": "^22.5.4",
     "@types/react": "^18.2.64",
     "@types/react-dom": "^18.2.21",
-    "@typescript-eslint/eslint-plugin": "^7.1.1",
-    "@typescript-eslint/parser": "^7.1.1",
     "@vitejs/plugin-react": "^4.2.1",
     "@vitest/ui": "^2.1.5",
     "dotenv": "^16.4.5",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
     "jsdom": "^25.0.1",
     "rollup-plugin-visualizer": "^5.14.0",
     "typescript": "^5.2.2",
