+++ frontend/src/hooks/useAuth.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { AuthContext } from "./AuthContext";
+
+export function useAuth() {
+    return useContext(AuthContext);
+}
