+++ frontend/src/routes/auth/signout.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from "@/hooks/Auth";
+import { useAuth } from '@/hooks/useAuth';
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
@@ -12,13 +12,13 @@ export default function SignoutPage() {
             signOut();
             setStatus("You have been signed out! Redirecting momentarily...");
         }
-    }, []);
+    }, [session, signOut]);
 
     useEffect(() => {
         if(session === null) {
             navigate("/");
         }
-    }, [session]);
+    }, [session, navigate]);
 
     return (
         <h1>{status}</h1>
