+++ frontend/src/routes/root.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from "@/hooks/Auth";
+import { useAuth } from '@/hooks/useAuth';
 import { Link } from "react-router-dom";
 
 export default function Root() {
