--- frontend/tests/vitest/__mocks__/Auth.ts
@@ -1,11 +1,0 @@
-vi.mock('@/hooks/Auth', () => ({
-    useAuth: vi.fn(() => ({
-        session: null,
-        loading: false,
-        signUp: vi.fn(),
-        login: vi.fn(),
-        sendPasswordReset: vi.fn(),
-        setNewPassword: vi.fn(),
-        signOut: vi.fn(),
-    })),
-}));
