+++ frontend/tests/vitest/setup.ts
@@ -3,7 +3,6 @@ import { cleanup } from '@testing-library/react';
 import * as matchers from '@testing-library/jest-dom/matchers';
 import '@testing-library/jest-dom';
 
-import './__mocks__/Auth';
 import './__mocks__/react-router-dom';
 
 expect.extend(matchers);
